problems: not working for infinite case yet, and not working on own examples

+++ timer.go
@@ -6,7 +6,7 @@ import (
 )
 
 func setTimer(f float64, c chan<- struct{}) {
-	timer := time.After(time.Duration(f) * time.Second)
+	timer := time.After(time.Duration(f * float64(time.Second)))
 	go wait(timer, c)
 }
